Avoid repetition of code
